Fixed stack overflow on DealerView
mkBetGraph will recurse forever if the list of fibonacci subsets is not
sorted.  We can't rely on sqlite to sort the entities by Value because
they are stored as strings (so "F34" comes before "F5"), therefore we
sort the results of getScrumBets manually before passing to mkBetGraph.

mkBetGraph is a time bomb as it is, but I'll deal with that later.